Change to use Graphics2D and cache graphics objects when possible.
We now use double-precision rendering with Shape objects. I also changed the
placement and size of some objects. If tile size hasn't changed since the last
time the control was rendered, the shapes should be cached and not generated
again. Further improvements---rendering to an image, caching that, and
rendering it to the screen, or caching the shapes statically, for
instance---should be possible.