Add a hierarchy of classes to replace numeric/legacy events.
These are not supported in the XML or elsewhere in the code; when they are,
they will be tags inside the tiles, replacing the numeric attributes. Once
exploration uses explorer DC, we can let that decide whether to update the
player maps with them rather than removing them by hand or always not adding
them. I plan to write a small program to convert numeric events in a map to
these, though precisely how will be somewhat tricky. And I intend to update
player maps with the events of this kind they have previously discovered.

NothingEvents should only be used to make these non-nullable, and omitted in
output.

It'd be nice to have an optional (XML) attribute (represented in the model and
UI, obviously) showing whether the player had investigated a battlefield, cave,
town, etc.

Next steps:
1. Model support: A tile can have one of these. (At most one, or any number?)
2. XML (tag) support. Need to decide whether to use one tag total or one tag
per event type.
3. UI support.