Change a Tile's TileType to be immutable.
This is because static analysis is warning me of too many methods in Tile
already, and I'm about to add more. Refactoring so Units, Fortresses, and (to
be added next) Events are all handled in one collection would alleviate that
problem significantly.

This change required me to change the terrain-type-changing code so that we're
changing to a new Tile behind the GUITile.