A new custom exception for illegal XML constructs (like rivers outside the map).
This is the first commit of yet another XML-map-reading attempt. This time, I'm
trying to avoid mixing reading data in with checking that it makes sense for
our model so much; instead, we'll read the XML into a tree of nodes, which will
then validate themselves and turn themselves into a map. Some of what I plan to
do in this version will also make adding dynamic data types (custom units,
etc., from file ...) much easier.