A factory for Nodes. We look in a Map to determine what category a tag is, then based on that construct a Node for it.
If a single Node type is used by multiple tags with different minimum
attributes---all rivers have to have a direction, for example, except that
lakes are treated as a special case of rivers and don't have to have a
direction---the category-to-Node step is where we do the extra setup required.
After the Factory returns an object, what XML was behind it is a black box. On
the other hand, while I haven't done this yet, it should be possible to have a
single tag map to multiple categories (which may or may not be a good thing to
allow).

The other benefit of this Map- and Factory-based approach, as opposed to the
previous model where enum values knew what tags they represented, is that it
lets us define new tags within existing categories at runtime, or at least
elsewhere in the project.

Of course, all this is a poor-man's substitute for true algebraic data types,
like Haskell (which I'm trying to learn) or Ceylon (which I am eagerly
awaiting) ...