The entry point into the module, which the application will call.
Instead of special-casing every kind of tag with its own parsing method, which
recursively called the methods for its child tags, here we have a stack and use
it to build up a tree in memory---but more cheaply (es*pecially* in memory!
unless we make a mistake) than a DOM-based parser, since most of the stuff a
general parser has to care about (namespaces, end tags, etc.) is irrelevant to
our needs. We delay sanity-checking until the end, but then each Node knows its
own requirements and checks them and builds up the object it represents itself.