Move the DetailPanel to below the map, and reconfigure it to suit.
I'm doing this partly because the two-column arrangement, with one column for
the explanatory labels, was not a good use of space, but also because I plan to
use the right-hand column for *some*thing in the near future. I'd start with a
mini-map, if I could figure out how to make that work.