Make MapComponent always use a helper.
Since this is at least not much worse in terms of performance than the
alternative (even comparing it to routines that had less to draw!), if not
better (!), and simplifies the class considerably, we'll just go ahead and do
it. Though I'm not going to push this immediately, in hopes of receiving some
performance-test results.