Remove getTile() from MapGUI interface.
Since TileSelectionListener is the only user of that method, and other
implementations will provide their own way of letting users change the selected
tile, and the method's signature exposes an implementation detail (the GUITile
type), we'll just remove it and make TileSelectionListener require a MapPanel
rather than any implementer of the interface. Though it could probably be
modified to remove that requirement easily enough.