Add and use new SystemOut wrapper around System.out
FindBugs still insists that System.out is always null. We had been using a
wrapper method in one place and importing two FindBugs JARs to tell it to
ignore it there, but here we add a new class that wraps System.out, and then
use it, which lets us drop the FindBugs JARs from the classpath.