Replace nearly all uses of MineralKind with String, and deprecate the enum.
We want to be able to handle arbitrary kinds of minerals without adding them to
the enum; there's no reason why *not* to just make it a string type, it now
seems to me. But LegacyTable (part of the Mark II encounter system, which will
go away once this conversion is finished) still needs it to create the table,
so we won't get rid of the enum entirely (or modify it).