Add a Set implementation backed by an array.
There are several places in the code where I *really* don't want the
performance hit a HashSet gives me (and TreeSet has its own problems, as I
discovered a while back and documented in the relevant comments), and would
rather have the fast performance of an array, but still occasionally rely on
the Set's guarantees. This implementation should be blazing fast for iteration
but a little slower (than a stock ArrayList---I don't know about HashSet) on
insertion.