Resize labels to show entire text at fixed width.
This is the problem that has been vexing me for weeks; I finally found a
solution on the Internet
<http://blog.nobel-joergensen.com/2009/01/18/changing-preferred-size-of-a-html-jlabel/>
and adapted it here. The Tree doesn't quite work (since there's no way to get
the available width *at this depth*) but except for being a possibly easier
base for unit details and such it's far inferior to the List implementation
(which is even typesafe!) I'll be removing the Tree and Table implementations
shortly, to keep code complexity down.