Make PlayerCollection just an Iterable<Player>, not a Collection anymore.
This makes both the "return of collection" warning in SPMap go away (though it
still should apply) and fixes the "too many methods" warning in
PlayerCollection. One method from Collection wasn't removed, but will be
changed to better fit the class rather than the Collection interface in the
next commit.