Automatically select appropriate item in a list.
When populating a list, automatically select the first item; when adding a new
item at the user's request (i.e. a new Job or Skill), select it.

This is the third or fourth attempt at this; every previous attempt ended with
the GUI routinely crashing, probably because of Swing concurrency issues. This
implementation doesn't suffer from those issues, and is additionally far less
intrusive, so a superior attempt anyway.