Fix inter-window and inter-JVM drag-and-drop to actually work.
What I'd tried to do, using the "drop target context"'s parent component,
would never have worked because the drop target context is the component on
which this decision is being made. Instead, we look at the "source", which
either isn't a component at all, or is the window, or something in inter-JVM
transfers.