Redesign DetailPanelNG.
This first replaces the TileDetailPanel with a mention of the tile's
coordinates in the header label above the list (since the other piece of
information in the TileDetailPanel is duplicated in the contents of the list).
Then, since there are only two top-level components, it makes the
DetailPanelNG a JSplitPane subclass instead of doing the layout with manual
layout hacking that was quite brittle---this is basically what I've always
wanted, except that it's not possible to "undock" the panels.