Fix the catch-22 of constructor access warnings by replacing named inner classes with anonymous objects.
Instantiating an inner class whose constructor is private (or not present, if the class is private) triggers a warning, because
the compiler has to raise its visibility level to make that work. Having a private class, inner or not, with a public constructor
(or any other public member) also triggers a warning. And CodePro (which I'm finally able to use reliably again) warns about any
member (including a constructor) with unspecified (i.e. package-private) access level. So in this case, we simply drop the whole
thing and make the objects anonymous objects rather than named classes.