In ProxyJob/Skill, don't assume add() used ours.
Usually, addJob() and addSkill() simply add the object they are passed
as a parameter to their containing set. However, ProxyJob is a
collection of ProxySkills, and ProxyWorker is a collection of ProxyJobs,
so adding a Job to a ProxyWorker or a Skill to a ProxyJob nominally
succeeds (the Worker has a Job by that name, or the Job a Skill, now),
but the object is discarded, and so reusing it (or relying on
modifications made to it before-hand, since only its name is used) is a
leak.

This fixes the egregiously annoying bug, present from the introduction
of the multi-map-experience-adding feature, that workers didn't get
credit for hours in skills they didn't already have.

It's quite possible that a better way to fix this, at least in the case
of adding skills, would be to make ProxyJob use the Skill's state. But
this will do for now.

The test case that helped track this down will follow in the next
commit.