Map viewer: Pass choice-failure to caller
Previously, even after AppChooser was modified to not quit
as soon as an app was selected but only after it
*successfully* started, selecting the map viewer always
caused the AppChooser to quit even if the user canceled out
of the file-selection dialog, because a failure to choose a
file was logged but otherwise handled by a "normal" return.

Instead, the ChoiceInterruptedException is now wrapped in a
DriverFailedException and thrown up the stack.