ImageLoader: Fix potential NPE
ImageIO#read() is documented to return null if no ImageReader is found that
claims to be able to read the image in the stream. I've never seen this happen,
but this means that the "the returned value may be null" warnings from Eclipse
were not just a result of the JDK not having proper nullity annotations, so we
fix this potential NPE properly here by turning a null return value into a
thrown IOException.