WorkerTreeModelAlt: Handle proxies specially in moving workers
This is very much a hack, that only works when the member and both old and new
owners are proxies and all proxies have the same number of proxied objects, but
when those conditions hold, loop through the proxied objects and do the
corresponding move instead of adding a proxy member to (because of the
implementation of ProxyUnit.addMember) all proxied units.

This prevents a nasty map-corrupting crash in the XML-writing code when it
tried to find a writer for a ProxyWorker.