SPMapNG.addFixture: Detect and try to handle duplicate IDs
If the new fixture does not have an ID that duplicates one already on the tile,
add it as usual. If it does, and the existing one either equals, is a subset of,
or is a superset of the new fixture, remove the old one and add the new one.
Otherwise, add it as usual but log an exception (to get the stack trace) and what
kind of fixtures both are.

This appears to finally fix the duplicate-HQ problem, so long as maps don't
diverge too much.