Set up Eclipse null-annotations; fix fallout.
Set up Eclipse classpath to use nullity annotations (stored
in annotations/), and add annotations for Map.get() and a
few others that came up in fixing the fallout from that.

Now that Eclipse knows that Map.get() can return null, it
objects even to cases where we check that the key is in the
map before calling it, so the bulk of this commit is adding
NullCleaner calls around the objects returned by get() in
such situations. But instead, maybe we should return to the
pattern of calling get() unconditionally, then checking
whether it returned null, in most cases.