StreamingLabel: Extend JEditorPane instead of JLabel
This makes the text in the label selectable, makes it start
at the top instead of in the middle and then moving the
start up as more text accumulates, and makes the component
report an accurate size so scrollbars will automatically
work if it's surrounded by a JScrollPane.

However, this by default has a larger spacing between
paragraphs than we like for the ResourceAddingFrame, so we
fix that in the HTML we pass in there.