ResourceAddingFrame: Listen for Enter only, not item changes
JComboBox calls on its ActionListeners whenever the selected
item changes. As there are other listener interfaces for
that, this is *never* what I want. The way around this that
I found is to get the JTextField that the ComboBoxEditor
inside the JComboBox uses for letting you enter the text,
and add an ActionListener to *that*. Unfortunately, that
text field isn't guaranteed by the type system to be a
JTextField, so this might break on some setups. I've only
tested it on Mac.