Make Forests have unique-per-instance ID numbers.
Unlike every other fixture type so far, if a Forest is read
from XML without an ID, it is *not* set to the first unused
ID, but rather it defaults to -1 (the old constant ID #).
However, the EchoDriver now sets all "primary" forests' ID
numbers to the first number above the current last used ID
number plus the number of its tile, counting across and then
down from the origin.