ExplorationModel: Rudimentary moved-unit-ghost fixing
When moving, if, in a nearby tile, a submap has a fixture
that isn't there in the main map but is also somewhere else
in the same submap, remove it.

This does very little to solve the problem posed by units
far from a player's HQ (or sworn villages) being recorded in
his or her map, then moving in the main map: recording their
current position would be an information leak to the player,
as would unconditionally removing them. But it's a start.
And it will have to do until we have a proper "fog of war"
feature.