ProxyUnit: Fix removeMember() to work when member is a proxy
When the member is a proxy, and doesn't equal (==) any
member of one of our proxied units (which is the most common
case when the user uses the Dismiss pop-up menu item in the
worker-management GUI when he specified more than one map on
the command line), test them against each unit member
proxied by the proxy "member" that was passed in until one
matches.

In addition, add logging statements so we know "immediately"
(instead of by inspecting the XML when we save the map file)
if something still resulted in no change, and a test of this
functionality.