YAMapReader: Memoize which readeers and writers to use for child tags.
That is, keep caches mapping tags to readers and classes to writers,
so we don't have to iterate over the possibilities every time. While
this may hinder performance for the very smallest maps slightly (e.g.
those common in the unit tests), it should speed up reading and
writing significantly in more usual cases, and increasingly so as
map size increases.