ExplorationCLIHelper: Add a pathfind-toward-given-point option
This is intended for the case of purposeful movement rather
than exploration. It uses Dijkstra's algorithm, as provided
by the pathfinder object, but (unless no subordinate maps
are provided) bases it on the first subordinate map, so I
think it will tend to avoid unexplored areas. (And it won't
change the route if new discoveries reveal that the planned
route isn't optimal.)