SPFormatException: Take row and column as Integers instead of taking Location as a parameter and field.
This way, when we remove the native("jvm") header from the
module as a whole, this abstract base class can remain as-is
without needing to be made JVM-only. Most of its subclasses,
which have to take classes from the JDK's XML interfaces as
their parameters, unfortunately will need native headers.