Add lovelace.util.common::simpleMap and use throughout.
If the same problem as with ceylon.language::set applies, using
ceylon.language::map with the declarative idiom will cause the
compiler to create an extra inner class for each caller. I made
this wrapper to avoid that side effect, and also used it to
replace call-sites where I'd wrapped the Entries in a Tuple to
avoid that effect.