Remove support for sandbars in the map.
There are (were) over 3000 of them in the world map, many in
locations that don't really make sense, and I've never had any
ideas of any game-world effects they might have, unless they
happened to be _in_ a river, which the model doesn't support.

If the XML-reading code encounters the tag, it will trigger a
warning, but pass on by (unless warnings have been made fatal).
The code for reading from a database will also comment on but
safely ignore mentions of sandbars in the table where older
versions of the code would have written them.