TerrainChangingMenu: Fix toggling items to actually work.
For some reason, when we use the button models as our base, they
get set to 'unselected' unconditionally (_not_ in my code), then
when clicked get set internally to 'selected' before entering the
action-listener method, so it assumes that the user is trying to
remove a river or mountain. So I've now moved back to inverting
the actual state in the map.