TurnRunningCLI: Operate on proxies, not just units in main map.
To be able to use ExplorationCLIHelper for the exploration applet our model
has to be an IExplorationModel, but ExplorationModel's getUnits() method
only goes to the main map. So this commit copies the algorithm from
WorkerModel into TurnRunningCLI and uses it instead. We make
sure not to set a ProxyUnit as the selected unit in the
driver model, however, as the text "proxy for" would leak
out in the UI and the model would fail to find it, making
its nominal location (-1, -1).