Replace ceylon.http.server dependency with Takes.
This reduces the size of the final jar by about one megabyte. This is
something of a compromise; the first lightweight HTTP server I found was
[JLHTTP](https://www.freeutils.net/source/jlhttp/), but as it's GPL-2 and this
project is GPL-3 its license is incompatible. The second I tried was
[NanoHTTPD](https://github.com/NanoHttpd/nanohttpd), but that triggered
eclipse/ceylon-ide-intellij#669 in the IDE and a 'duplicate class' backend
error (reported as being in an unrelated module) in the command-line compiler,
so I had to give up on that. Because this pulls in dependencies too (just
slightly lighter-weight ones that ceylon.http.server), I should try NanoHTTPD
again when it makes its next release.