skillAdvancementPanel: Add late annotation to listeners list.
Without this, it never actually gets initialized, causing an NPE
when the class tries to add a listener to the list. This is clearly
a bug in the runtime, which I need to report if I can distill a MWE,
but I have no idea why it's only surfacing now and why it didn't
occur when I committed 90ae8ec942c5c7e507f40e9ba8e8e01c4b37c66e.