PathWrapper: Drop ref from doc and add native("js") body.
The compiler bundled in the IntelliJ plugin sometimes (seemingly often)
complained that the reference to ceylon.file::Path in the todo() string
of the native-header class was a reference to a nonexistent type and
that PathWrapper couldn't compile under native("js"). So I changed the
string to not be a link to the preferred type and added an explicitly
native("js") implementation (which should be byte-for-byte identical
when compiled, I would expect).