TurnRunningCLI: 'Taking provisions' feature. Not tested.
In the 'explore' applet, when a unit leaves a fortress, the user
is now asked if the unit should take provisions, and if so then
asked to choose them from the fortress's resources.

The point of this is for the food-consumption mechanic to
eventually be able to limit consumption to food in the same
unit or fortress.