TypesafeXMLReader: Disable external entity support.
I'm not sure this mitigation actually worked (because of the way Ceylon
doesn't translate its quasi-primitives into their Java equivalents when
calling a Java SDK method with argument type 'Object'). And the security
impact of the vulnerability is minimal: it would require an attacker to
trick the user into opening a specially crafted file. But if I ever make
a Web-based version of any of the assistive programs, this would become
seriously important, so it's best to fix it now.