IDriverUsage et al: Drop multiple-invocation support.
As a holdover from the days when apps where chosen using _options_,
possibly multiple apps per execution, the IDriverUsage specified a
nonempty Iterable of Strings that could be used to invoke the app. This
changes that to a single String that is used to invoke it.

As it happened, no app supported more than one invocation anyway, so
this was just a matter of dropping the brackets around each string from
the source code; the more major change was fixing the code in the app
chooser that consumed this field.