PathfinderImpl: Fix 'is unvisited' calculation.
When I added caching to the implementation of Dijkstra's algorithm, I introduced a bug that
caused it to consider a point unvisited if *the start point* was unvisited, not if *that point*
was unvisited. This made the algorithm always abort if the destination was not adjacent to or
identical to the start tile.