ScrollAdjustmentListener: Use a field, not the model's valueIsAdjusting, as a mutex.
Previously we skipped handling a scroll event if the scrollbar declared itself
as 'currently adjusting', to avoid stack overflows from setting the scrollbar
models in response to setting the cursor in response to changes in the
scrollbar position. Unfortunately this prevented us from moving the cursor
when the user was dragging the scrollbar, so we now use a Boolean field
to guard the critical section, and skip the event handling only when it
is set.