Add fromSecondMap() to IMultiMapModel to remove mutable-map-ref from IOHandler.
IOHandler needs to be able to open the second map in a map viewer window, but
before this commit the best way to do that was to create a mutable-map-and-filename
entry and use the from-entry constructor of ViewerModel. Now we can get the
second map and its filename as a driver model and use ViewerModel's copy
constructor.