ExplorationModel: Fix copying of noticed fixtures to submaps.
We were mistakenly looking for a fixture in the main map that equaled
the given fixture's ID, instead of one whose ID equaled the given
fixture's ID. So of course nothing matched, and so nothing was ever
copied to submaps. Oops.

Bug introduced in a8d2fb1c7291186548c20c4548616c7b7659ceb4.