Rework database I/O.
Instead of reading child items from the database in a separate pass to make sure
they are read after their parents, read everything in one pass and store them in
a multimap by parent ID, then connect children to parents afterward.

This is necessary to support worker mounts and equipment, but significantly
simplifies the architecture of the DB I/O module as well, so we'll want to keep
this even if for some reason we end up throwing out the (not yet committed)
worker mounts/equipment changes.