Convert CI from Travis to Github Actions. (#86)
And make the CI build more robust in the absence of modules.ceylon-lang.org, since that just recently went 
down and the Maven interop (even with a local Maven repository cache, since the CI-installable version of 
the compiler is missing the HTTPS-for-Maven-Central patch) seems to have gone with it.

We now specify to use 64-bit Launch4J (which wasn't available when I first set up CI builds, or I would have 
used it!) which lets us drop the i386 `libc` dep, which is not found in the image used in the GitHub workflow.

When Ceylon Herd is down, we download a copy of `.ceylon/cache` (made on my laptop) from either the 
Actions cache or Google Drive. And because the only binary-available versions of the Ceylon compiler
no longer work with Maven Central (it requires HTTPS, and they try to use HTTP) we explicitly
download to the local Maven cache and `ceylon import-jar` all the Maven dependencies we directly or
indirectly use. This requires module descriptors for a lot of modules, and some have to be imported with
`--force` at first because of circular dependencies, but I did finally get it to work.

Merged as a squash commit because the details of how we got here are ugly in the extreme, and this also
needs to go on all currently-active branches.