Clear modified flag at end of map-reading.
With the modified flag now stored in the SPMapNG class and set by all
its mutators, it is invariably set to 'true' by the end of the
map-reading process, in both the database I/O package and both
XML-reading suites.  This change adds a statement to set it back to
'false' before returning the map to the caller in each place a map is
constructed from user input.