Port IFortress and IMutableFortress back to Java.
IFortress contained equals() and hashCode() defaults, which isn't
possible in Java (Ceylon lets an interface override methods from
Object); I coverted the Ceylon code back to Java, but commented
it out so the files will compile. Once the concrete implementation
is ported the code will need to be moved there.