Port TurnApplet, TurnAppletFactory, and AbstractTurnApplet to Java.
There's one method where the more compact stream-based implementation failed
to compile with error messages that were clearly themselves in error, so
I commented that out and wrote a basically-equivalent (and slightly superior
as far as full correctness is concerned) loop-and-if-based implementation.
Need to investigate once I'm working in an IDE again.