Rework database I/O.
Instead of reading child items from the database in a separate pass to
make sure they are read after their parents, read everything in one pass
and store them in a (poor-man's-)multimap by parent ID, then connect
children to parents afterward.

This is necessary to support worker mounts and equipment, but
significantly simplifies the architecture of the DB I/O module as well,
so we'll want to keep this even if for some reason we end up throwing
out the (not yet committed) worker mounts/equipment changes.

"Cherry-picked" (manually ported) from
58513032010fb5dd8216df2ac126bd0e66693c3d, except methods now unused
(and the entire DBMemoizer class) are removed instead of merely being
marked deprecated, and a couple of lines that should have been in the
following commit weren't included here.