Throughout, replace MalformedXMLException with XMLStreamException.
MalformedXMLException was introduced when we ported to Ceylon, to be able to have the
thrown-or-returned-exceptions part of the interface be 'pure Ceylon', not depending on
the Java SDK or requiring consumers to be annotated 'native'. Now that we're back in
Java, there's no point in wrapping the XMLStreamException instances in an exception
with identical semantics.