Start on SPMenu.forWindow() conversion.
I want to move away from requiring the SPMenu factory method to iterate parents until it
finds a frame, since that's slow and not guaranteed (by the type system) to succeed.
This commit adds the necessary methods in the menu classes and converts the callers where the
window isn't adding a menu to itself; those need more thought, because the reason we went
the forWindowContaining() route in the first place was to avoid 'this' escaping
constructors.