Use multiline strings in more places.
Mostly to avoid having to backslash-escape double-quote characters. (This commit also
includes a few cases where I simply changed the string to not include double-quotes,
so as to avoid the need to convert. And when a converted string had been the input
to String.format(), I changed that to use String::formatted().