IDEA config: Disable 'public constructor in non-public class' warning.
It mostly triggers in code where our choices are public (triggering this warning), protected (triggering
'protected in final class' or similar warnings), private (causing synthetic methods to be generated,
triggering warnings and also allowing bypass of that access control), and package-private (triggering
warnings because no explicit access keyword is specified). 'public' is the least-worst option, so just
disable that warning.